Fix bug test load gm.

+++ src/ringmesh/io_full_geo_model.cpp
@@ -187,6 +187,23 @@ namespace {
 
     }
 
+    bool save_mesh(
+        const GeoModelMeshElement& geo_model_element_mesh,
+        const std::string& name,
+        const GEO::MeshIOFlags& flags )
+    {
+        if( geo_model_element_mesh.type() == GME::REGION ) {
+            const Region& region = geo_model_element_mesh.model().region(
+                geo_model_element_mesh.index() ) ;
+            if( !region.is_meshed() ) {
+                // a region is not necessary meshed.
+                return false ;
+            }
+        }
+        GEO::mesh_save( geo_model_element_mesh.mesh(), name, flags ) ;
+        return true ;
+    }
+
     /************************************************************************/
     class AsterIOHandler: public GeoModelIOHandler {
     public:
@@ -415,11 +432,13 @@ namespace {
         logger->set_quiet( true ) ;
         GEO::MeshIOFlags flags ;
 //        flags.set_attribute( GEO::MESH_ALL_ATTRIBUTES ) ;
-        GEO::mesh_save( geo_model_element_mesh.mesh(), name, flags ) ;
+        bool is_saved = save_mesh( geo_model_element_mesh, name, flags ) ;
         logger->set_quiet( false ) ;
 
-        zip_file( zf, name ) ;
-        GEO::FileSystem::delete_file( name ) ;
+        if( is_saved ) {
+            zip_file( zf, name ) ;
+            GEO::FileSystem::delete_file( name ) ;
+        }
 
 //        std::string root_name ;
 //        std::string handler ;
